backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
requests
algoliasearch
pydantic[email]
